Issue #28594 Update adjustinvalue docnum to ordnum
In the original commit in this change, we added the document number parameter
as an optional parameter and saved it to the invhist_docnumber field.  Turns
out this is not consistent with the way other adjustments are handled and
other issues complaining about docnumber not showing on the invhist report
fix it by moving the value to invhist_ordnumber.  So now in adjustinvvalue
we update the function to match the value handling of public.invadjustment.

Note that we now set invhist_ordtype to "AD".  This was not part of the
original functionality, but should be a non-breaking change: this change is
also consistent with the public.invadjustment function, including in cases
where no document number is provided (it is an optional field).