Issue #28594 Remove unused adjustinvvalue var
The _glreturn variable was just a throwaway destination for the select
return value.  While that value would be a nice return value in some
way/shape/form, we're already returning the invhist_id and there is no
real norm in xTuple ERP (that I know of) for returning compoundtypes (or
similar a la JSON, hstore, etc).