issue #30139 - fixSerial as part of upgrade process
The bug as reported was a duplicate key on insert error on potype. The
real problem was that the sequence for potype_id was being created
with a lower number than the max potype_id as a result of data
migration. The code change in this PR tries to prevent this _kind_
of problem from arising in the future.